backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
google-generativeai>=0.3.1
python-dotenv>=1.0.0
httpx>=0.25.0




